Changed a logic in RemoteAuthorizationProvider, RemoteOrdersDataProvider. Data for request generates in class ParserToJSON. Created class AuthorizationDataParser.

+++ LV-116.iOS.SS.Restaurant/LV-116.iOS.SS.Restaurant/Providers/OrdersDataProvider/OrdersDataProvider.h
@@ -19,6 +19,6 @@
 + (void)deleteTableOrderWithOrderId:(int)orderId responseBlock:(void (^)(NSError*))callback;
 
 
-+ (void)postTableOrderWithTableModel:(int)tableId responseBlock:(void (^)(NSUInteger,  NSError*))callback;
++ (void)postTableOrderWithTableModel:(NSInteger)tableId responseBlock:(void (^)(NSUInteger,  NSError*))callback;
 
 @end
